// src/app/productos/interface/producto.interfaces.ts
export interface Producto {
    id:     string;
    nombre: string;
    creador: string;
    precio: string;
    tipo:   string;
    descripcion: string;
    alt_img: string;
    cantidad: string;
}

// export enum Tipos {
//     cerca= "cerca",
//     salon= "salón",
//     escena= "escena",
//     grandes= "grandes ilusiones",
//     libro= "libro",
//     dvd= "dvd",
//     complementos= "complementos"
    
// }